+++ lib/platform_interface.dart
@@ -113,8 +113,7 @@ class WebResourceError {
     this.domain,
     this.errorType,
     this.failingUrl,
-  })  : assert(errorCode != null),
-        assert(description != null);
+  });
 
   /// Raw code of the error from the respective platform.
   ///
@@ -370,12 +369,13 @@ class WebSetting<T> {
   @override
   bool operator ==(Object other) {
     if (other.runtimeType != runtimeType) return false;
+    // ignore: test_types_in_equals
     final WebSetting<T> typedOther = other as WebSetting<T>;
     return typedOther.isPresent == isPresent && typedOther._value == _value;
   }
 
   @override
-  int get hashCode => hashValues(_value, isPresent);
+  int get hashCode => Object.hash(_value, isPresent);
 }
 
 /// Settings for configuring a WebViewPlatform.
@@ -397,7 +397,7 @@ class WebSettings {
     this.gestureNavigationEnabled,
     this.allowsInlineMediaPlayback,
     required this.userAgent,
-  }) : assert(userAgent != null);
+  });
 
   /// The JavaScript execution mode to be used by the webview.
   final JavascriptMode? javascriptMode;
@@ -455,7 +455,7 @@ class CreationParams {
     this.userAgent,
     this.autoMediaPlaybackPolicy =
         AutoMediaPlaybackPolicy.require_user_action_for_all_media_types,
-  }) : assert(autoMediaPlaybackPolicy != null);
+  });
 
   /// The initialUrl to load in the webview.
   ///