+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 5.0.3
+Fixed bug to properly show plugin version
+
+## 5.0.2
+Added new functionalities to the plugin: 
+*deeplink tracking
+*user matching
+Updated internal dependencies.
+
 ## 5.0.1
 Changed a way how plugin can be initializes for Android. There is no need to initialize plugin from native Java or Kotlin code anymore.  This all can be done from the flutter side.
 ## 5.0.0